refactor to better see import times
