+++ python/cudf/cudf/core/udf/__init__.py
@@ -1,9 +1,21 @@
 # Copyright (c) 2022-2023, NVIDIA CORPORATION.
-from . import (
-    groupby_lowering,
-    groupby_typing,
-    masked_lowering,
-    masked_typing,
-    strings_lowering,
-    strings_typing,
-)
+#from . import (
+#    groupby_lowering,
+#    groupby_typing,
+#    masked_lowering,
+#    masked_typing,
+#    strings_lowering,
+#    strings_typing,
+#)
+
+from . import strings_typing, strings_lowering
+strings_typing.register_strings_typing()
+strings_lowering.register_strings_lowering()
+
+from . import masked_typing, masked_lowering
+masked_typing.register_masked_typing()
+masked_lowering.register_masked_lowering()
+
+from . import groupby_typing, groupby_lowering
+groupby_typing.register_groupby_typing()
+groupby_lowering.register_groupby_lowering()
