+++ python/cudf/cudf/core/udf/utils.py
@@ -38,6 +38,7 @@ from cudf.utils.dtypes import (
     STRING_TYPES,
     TIMEDELTA_TYPES,
 )
+from cudf.core.udf._compile_udf import compile_udf, make_cache_key
 
 if TYPE_CHECKING:
     from collections.abc import Callable
@@ -63,11 +64,6 @@ MASK_BITSIZE = SIZE_TYPE_DTYPE.itemsize * 8
 
 precompiled: cachetools.LRUCache = cachetools.LRUCache(maxsize=32)
 
-# This cache is keyed on the (signature, code, closure variables) of UDFs, so
-# it can hit for distinct functions that are similar. The lru_cache wrapping
-# compile_udf misses for these similar functions, but doesn't need to serialize
-# closure variables to check for a hit.
-_udf_code_cache: cachetools.LRUCache = cachetools.LRUCache(maxsize=32)
 
 
 UDF_SHIM_FILE = os.path.join(
@@ -143,75 +139,6 @@ def _mask_get(mask, pos):
     """Return the validity of mask[pos] as a word."""
     return (mask[pos // MASK_BITSIZE] >> (pos % MASK_BITSIZE)) & 1
 
-
-def make_cache_key(udf, sig):
-    """
-    Build a cache key for a user defined function. Used to avoid
-    recompiling the same function for the same set of types
-    """
-    codebytes = udf.__code__.co_code
-    constants = udf.__code__.co_consts
-    names = udf.__code__.co_names
-
-    if udf.__closure__ is not None:
-        cvars = tuple(x.cell_contents for x in udf.__closure__)
-        cvarbytes = dumps(cvars)
-    else:
-        cvarbytes = b""
-
-    return names, constants, codebytes, cvarbytes, sig
-
-
-def compile_udf(udf, type_signature):
-    """Compile ``udf`` with `numba`
-
-    Compile a python callable function ``udf`` with
-    `numba.cuda.compile_ptx_for_current_device(device=True)` using
-    ``type_signature`` into CUDA PTX together with the generated output type.
-
-    The output is expected to be passed to the PTX parser in `libcudf`
-    to generate a CUDA device function to be inlined into CUDA kernels,
-    compiled at runtime and launched.
-
-    Parameters
-    ----------
-    udf:
-      a python callable function
-
-    type_signature:
-      a tuple that specifies types of each of the input parameters of ``udf``.
-      The types should be one in `numba.types` and could be converted from
-      numpy types with `numba.numpy_support.from_dtype(...)`.
-
-    Returns
-    -------
-    ptx_code:
-      The compiled CUDA PTX
-
-    output_type:
-      An numpy type
-
-    """
-    key = make_cache_key(udf, type_signature)
-    res = _udf_code_cache.get(key)
-    if res:
-        return res
-
-    # We haven't compiled a function like this before, so need to fall back to
-    # compilation with Numba
-    ptx_code, return_type = cuda.compile_ptx_for_current_device(
-        udf, type_signature, device=True
-    )
-    if not isinstance(return_type, MaskedType):
-        output_type = numpy_support.as_dtype(return_type).type
-    else:
-        output_type = return_type
-
-    # Populate the cache for this function
-    res = (ptx_code, output_type)
-    _udf_code_cache[key] = res
-
-    return res
 
 
 def _generate_cache_key(frame, func: Callable, args, suffix="__APPLY_UDF"):
